move pydantic models into database.schema
